trunner: restructure parsing 3rd party harnesses
JIRA: PD-292
